# [GLUTEN-3141][CH] Rewrite date comparison expression into a more efficient way