## What changes were proposed in this pull request?

(Please fill in changes proposed in this fix)

Fixes: #3141

Test result is given as following.

data set
```sql
CREATE TABLE default.test_time_cmp (ts INT, val INT) USING parquet;
insert into test_time_cmp select 1693497600 + (rand() * 1000000000) % 2592000 as ts, id as val from range(0, 10000000);
```

performance comparison
| condition   |      before      |  after |
|----------|:-------------:|:------:|
| from_unixtime(ts, 'yyyy-MM-dd') > '2023-09-15' |   0.862 seconds | 0.367 seconds |
| from_unixtime(ts, 'yyyy-MM-dd') >= '2023-09-15' |    0.879 seconds  |   0.347 seconds |
| from_unixtime(ts, 'yyyy-MM-dd') < '2023-09-15' | 0.887 seconds |    0.334 seconds |
| from_unixtime(ts, 'yyyy-MM-dd') < '2023-09-15' |  0.919 seconds |     0.319 seconds |
| from_unixtime(ts, 'yyyy-MM-dd') = '2023-09-15'| 0.872 seconds |    0.371 seconds |
| to_date(from_unixtime(ts)) > '2023-09-15' | 4.507 seconds |     0.331 seconds |

## How was this patch tested?

(Please explain how this patch was tested. E.g. unit tests, integration tests, manual tests)
unit tests


(If this patch involves UI changes, please attach a screenshot; otherwise, remove this)